+++ src/main.rs
@@ -183,7 +183,7 @@ async fn main() -> std::io::Result<()> {
             .service(nodeinfo::get_nodeinfo_2_0)
             .service(nodeinfo::get_nodeinfo_2_1)
             .service(web_client::profile_page_redirect())
-            .service(web_client::post_page_redirect())
+            .service(web_client::post_page_overlay(&config))
             .service(
                 // Fallback for well-known paths
                 web::resource("/.well-known/{path}")